feat: addressed code review
